pyproject.toml
[tool.poetry]
name = "social_m_chdt"
version = "0.0.1"
description = "A Django project"
authors = ["Nico A <nicolasabitante@gmail.com.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2"
psycopg = {extras = ["binary"], version = "^3.1.8"}
djangorestframework = "3.15.2"
pytest = "^8.2.2"
pytest-django = "^4.5"
factory_boy = "^3.3.0"
coverage = "^7.5.4"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "social_med.settings"